feat(v0.3.0): hybrid storage + settings submenus + brightness slider
- SD = source of truth, NVS = cache. JSON sidecar pro session.
- LIST_SESSIONS/META_PUSH/DELETE_FILE/DELETE_JSON commands fuer SD-resync
  + delete-propagation NVS<->SD.
- Settings als kategorisierte submenue-architektur (saunameister, verbindung,
  zeit, anzeige, daten-export, diagnose, speicher, danger zone).
- Helligkeits-slider via LEDC PWM (5-100%, NVS-persistent).
- Manuelle zeit-einstellung mit DST-aware mktime + immer TZ env setzen.
- Test-session cleanup (peak<40C, atomar im model um event-flood zu vermeiden).
- One-shot legacy-fix routine fuer pre-v0.3 recovered sessions
  (TZ-korrektur + operator-mapping aus hardcoded liste).
- USB-CDC export auf RP2040 (?L, ?D) jetzt im README dokumentiert.
- ESP32 DRAM-headroom durch task-stacks + buffers nach PSRAM.

+++ SenseCAP_Indicator_ESP32/main/model/indicator_sensor.h
@@ -22,10 +22,10 @@ int indicator_sensor_get_data(struct view_data_sensor *out_data);
  *
  * @param cmd    Kommando-Typcode (z.B. SSC_CMD_SESSION_START)
  * @param data   Zeiger auf Payload oder NULL
- * @param len    Anzahl Payload-Bytes (max 30)
+ * @param len    Anzahl Payload-Bytes (max 510 in v0.3.0)
  * @return       >=0: Anzahl versendeter Bytes, <0: Fehler
  */
-int indicator_sensor_rp2040_cmd(uint8_t cmd, const void *data, uint8_t len);
+int indicator_sensor_rp2040_cmd(uint8_t cmd, const void *data, uint16_t len);
 
 #ifdef __cplusplus
 }
